Refactor Localization to *not* directly load the main and Research bundles
